refactor: Extract component from index file in specific files
