+++ packages/cozy-sharing/src/RefreshableSharings.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import SharingContext from './context'
+
+export const RefreshableSharings = ({ children }) => (
+  <SharingContext.Consumer>
+    {({ refresh }) =>
+      children({
+        refresh
+      })
+    }
+  </SharingContext.Consumer>
+)
