+++ packages/cozy-sharing/src/RefreshableSharings.spec.jsx
@@ -1,0 +1,43 @@
+import React from 'react'
+import { mount } from 'enzyme'
+
+import { createMockClient } from 'cozy-client'
+
+import AppLike from '../test/AppLike'
+import { RefreshableSharings } from './RefreshableSharings'
+import { SharingProvider } from './SharingProvider'
+
+const DumbComponent = () => {
+  return <div>test </div>
+}
+
+const AppWrapper = ({ children, client }) => {
+  return (
+    <AppLike client={client}>
+      <SharingProvider client={client}>{children}</SharingProvider>
+    </AppLike>
+  )
+}
+
+describe('RefreshableSharings', () => {
+  it('should test', () => {
+    const client = createMockClient({})
+
+    const component = mount(
+      <AppWrapper client={client}>
+        <RefreshableSharings>
+          {({ refresh }) => (
+            <DumbComponent
+              client={client}
+              t={x => x}
+              refreshSharings={refresh}
+            />
+          )}
+        </RefreshableSharings>
+      </AppWrapper>
+    )
+    const refreshMethod = component.find(DumbComponent).prop('refreshSharings')
+
+    expect(typeof refreshMethod).toBe('function')
+  })
+})
